requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.8.0
effdet==0.4.1
emoji==2.9.0
exceptiongroup==1.2.0
fastapi==0.105.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.2
greenlet==3.0.2
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.0
langdetect==1.0.9
langsmith==0.0.70
layoutparser==0.3.4
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==0.27.8
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.4
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
pikepdf==8.9.0
Pillow==10.1.0
portalocker==2.8.2
protobuf==4.25.1
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
pypdf==3.17.2
pypdfium2==4.25.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
qdrant-client==1.7.0
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.36.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
unstructured==0.11.2
unstructured-inference==0.7.15
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
